@@ -137,9 +137,9 @@ function run (client, done) {
     m: {
       s: 'g3',
       i: 3,
-      b: new Buffer([0xa, 0xb, 0xc])
+      b: Buffer.from([0xa, 0xb, 0xc])
     },
-    b: new Buffer([0xa, 0xb, 0xc]),
+    b: Buffer.from([0xa, 0xb, 0xc]),
     b2: new Uint8Array([0xa, 0xb, 0xc])
   }
 
