@@ -15,7 +15,6 @@
 // *****************************************************************************
 
 const path = require('path')
-const extend = require('util')._extend
 
 /**
  * @class Config
@@ -248,7 +247,7 @@ function Config (config) {
     systemPath: this.defaultLuaSystemPath()
   }
   if (typeof config.modlua === 'object') {
-    extend(this.modlua, config.modlua)
+    Object.assign(this.modlua, config.modlua)
   }
 
   /**
